+++ Sheeeeeeeeet/Views/ActionSheetItemCell.swift
@@ -16,4 +16,18 @@
 
 import UIKit
 
-open class ActionSheetItemCell: UITableViewCell {}
+open class ActionSheetItemCell: UITableViewCell {
+    
+    
+    // MARK: - Layout
+    
+    open override func layoutSubviews() {
+        super.layoutSubviews()
+        textLabel?.font = titleFont
+    }
+    
+    
+    // MARK: - Appearance Properties
+    
+    @objc public dynamic var titleFont: UIFont?
+}